Fix backend issues
